+++ models/Connection.cs
@@ -10,14 +10,14 @@ namespace LondonTube {
   class Connection {
     public Platform Source {get;}
     public Platform Target {get;}
-    public int standardTime {get;}
+    public Double standardTime {get;}
     public ModeType mode {get;}
 
     public TrackClosure Closure {get;set;}
 
     public Delay Delay {get;set;}
 
-    public Connection(Platform source, Platform target, int time, ModeType mode) {
+    public Connection(Platform source, Platform target, Double time, ModeType mode) {
       this.Source = source;
       this.Target = target;
       this.standardTime = time;
